refactor: split classes to new files
